pyproject.toml
[project]
name = "pypdf-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.68.0",
    "google-generativeai>=0.8.6",
    "pymupdf>=1.27.2.2",
]
